requirements.txt
streamlit>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python
pillow
numpy
matplotlib
seaborn
scikit-learn
pandas